refactoring loading dir, minor test update
