+++ src/teehr/utilities/generate_weights.py
@@ -11,8 +11,8 @@ import pandas as pd
 import dask
 import shapely
 
-from teehr.loading.utils_nwm import load_gdf
-import teehr.loading.const_nwm as const_nwm
+from teehr.loading.nwm22.utils_nwm import load_gdf
+import teehr.loading.nwm22.const_nwm as const_nwm
 
 
 @dask.delayed
