+++ tests/data/test_remote_list.csv
@@ -1,25 +1,25 @@
 filename
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t02z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t03z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t04z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t05z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t06z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t07z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t08z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t09z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t10z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t11z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t12z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t13z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t14z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t15z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t16z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t17z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t18z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t19z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t20z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t21z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t22z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230318/analysis_assim/nwm.t23z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230319/analysis_assim/nwm.t00z.analysis_assim.channel_rt.tm02.conus.nc
-gs://national-water-model/nwm.20230319/analysis_assim/nwm.t01z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t02z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t03z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t04z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t05z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t06z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t07z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t08z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t09z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t10z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t11z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t12z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t13z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t14z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t15z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t16z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t17z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t18z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t19z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t20z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t21z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t22z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230318/analysis_assim/nwm.t23z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230319/analysis_assim/nwm.t00z.analysis_assim.channel_rt.tm02.conus.nc
+gcs://national-water-model/nwm.20230319/analysis_assim/nwm.t01z.analysis_assim.channel_rt.tm02.conus.nc
